Edited to get correct output

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@
 	long long int Typex2Long_Int;
 	float TypeFloat;
 
-	printf("Size of a char: %ld byte(s)\n", sizeof(TypeChar));
-	printf("Size of an int: %ld bytes(s) \n", sizeof(TypeInt));
-	printf("Size of a long int: %ld bytes(s) \n", sizeof(TypeLongInt));
-	printf("Size of a long long int: %ld bytes(s) \n", sizeof(Typex2Long_Int));
-	printf("Size of a float: %ld byte(s) \n", sizeof(TypeFloat));
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(TypeChar));
+	printf("Size of an int: %lu bytes(s) \n", (unsigned long)sizeof(TypeInt));
+	printf("Size of a long int: %lu bytes(s) \n", (unsigned long)sizeof(TypeLongInt));
+	printf("Size of a long long int: %lu bytes(s) \n", (unsigned long)sizeof(Typex2Long_Int));
+	printf("Size of a float: %lu byte(s) \n", (unsigned long)sizeof(TypeFloat));
 	return (0);
 }
